--- pkg/tsdb/druid/response.go
@@ -1,218 +1,0 @@
-package druid
-
-import (
-	"sort"
-	"time"
-
-	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
-	"github.com/grafana/grafana-plugin-sdk-go/data"
-)
-
-type Framer interface {
-	Frame() *data.Frame
-}
-
-const timestampColumn = "timestamp"
-
-type TimeseriesResult []TimeseriesRecord
-
-// Frame returns data formatted as Grafana Frame.
-func (t *TimeseriesResult) Frame() *data.Frame {
-	columns := t.Columns()
-	fields := make([]*data.Field, len(columns))
-	for i, column := range columns {
-		labels := data.Labels{}
-		fields[i] = data.NewField(column, labels, t.Values(column))
-	}
-	return data.NewFrame("", fields...)
-}
-
-// Columns returns list of columns. It calls `Columns()` on first record. If
-// no records are available it returns nil.
-func (t *TimeseriesResult) Columns() []string {
-	for _, r := range *t {
-		return r.Columns()
-	}
-	return nil
-}
-
-// Values returns all values for given column.
-func (t *TimeseriesResult) Values(column string) interface{} {
-	if len(*t) == 0 {
-		return nil
-	}
-	results := make([]interface{}, len(*t))
-	for i, r := range *t {
-		results[i] = r.Value(column)
-	}
-	return toTypedResults(results)
-}
-
-type TimeseriesRecord struct {
-	Timestamp time.Time              `json:"timestamp"`
-	Result    map[string]interface{} `json:"result"`
-}
-
-// Columns returns list of columns for given record.
-// The first column will always be "timestamp" followed by other columns sorter
-// alphabetically.
-func (t *TimeseriesRecord) Columns() []string {
-	columns := make([]string, len(t.Result)+1)
-	columns[0] = timestampColumn
-	i := 1
-	for c := range t.Result {
-		columns[i] = c
-		i++
-	}
-	sort.Strings(columns[1:])
-	return columns
-}
-
-// Value returns value for given column.
-func (t *TimeseriesRecord) Value(column string) interface{} {
-	if column == timestampColumn {
-		return t.Timestamp
-	}
-	v, ok := t.Result[column]
-	if !ok {
-		return nil
-	}
-	return v
-}
-
-type TopNResult []TopNRecord
-
-// Frame returns data formatted as Grafana Frame.
-func (t *TopNResult) Frame() *data.Frame {
-	columns := t.Columns()
-	fields := make([]*data.Field, len(columns))
-	for i, column := range columns {
-		labels := data.Labels{}
-		fields[i] = data.NewField(column, labels, t.Values(column))
-	}
-	return data.NewFrame("", fields...)
-}
-
-// Columns returns list of columns. It calls `Columns()` on first record. If
-// no records are available it returns nil.
-func (t *TopNResult) Columns() []string {
-	for _, r := range *t {
-		return r.Columns()
-	}
-	return nil
-}
-
-// Values returns all values for given column.
-func (t *TopNResult) Values(column string) interface{} {
-	results := []interface{}{}
-	for _, r := range *t {
-		results = append(results, r.Values(column)...)
-	}
-	return toTypedResults(results)
-}
-
-type TopNRecord struct {
-	Timestamp time.Time                `json:"timestamp"`
-	Result    []map[string]interface{} `json:"result"`
-}
-
-// Columns returns list of columns for given record.
-// It assumes that every map from Result has the same columns, so it gets
-// the list from first item.
-// The first column will always be "timestamp" followed by other columns sorter
-// alphabetically.
-func (t *TopNRecord) Columns() []string {
-	for _, result := range t.Result {
-		columns := make([]string, len(result)+1)
-		columns[0] = timestampColumn
-		i := 1
-		for c := range result {
-			columns[i] = c
-			i++
-		}
-		sort.Strings(columns[1:])
-		return columns
-	}
-	return nil
-}
-
-// Value returns values for given column.
-func (t *TopNRecord) Values(column string) []interface{} {
-	values := []interface{}{}
-	for _, result := range t.Result {
-		if column == timestampColumn {
-			values = append(values, t.Timestamp)
-			continue
-		}
-		v, _ := result[column]
-		values = append(values, v)
-	}
-	return values
-}
-
-func toTypedResults(raw []interface{}) interface{} {
-	// TODO this is not perfect. Maybe we should get this from the query?
-	switch raw[0].(type) {
-	// TODO this can be done with generics, but we should wait until go is updated in this repo
-	case string:
-		results := make([]string, len(raw))
-		for i, v := range raw {
-			vv, _ := v.(string)
-			results[i] = vv
-		}
-		return results
-	case int:
-		results := make([]int, len(raw))
-		for i, v := range raw {
-			vv, _ := v.(int)
-			results[i] = vv
-		}
-		return results
-	case int32:
-		results := make([]int32, len(raw))
-		for i, v := range raw {
-			vv, _ := v.(int32)
-			results[i] = vv
-		}
-		return results
-	case int64:
-		results := make([]int64, len(raw))
-		for i, v := range raw {
-			vv, _ := v.(int64)
-			results[i] = vv
-		}
-		return results
-	case float32:
-		results := make([]float32, len(raw))
-		for i, v := range raw {
-			vv, _ := v.(float32)
-			results[i] = vv
-		}
-		return results
-	case float64:
-		results := make([]float64, len(raw))
-		for i, v := range raw {
-			vv, _ := v.(float64)
-			results[i] = vv
-		}
-		return results
-	case bool:
-		results := make([]bool, len(raw))
-		for i, v := range raw {
-			vv, _ := v.(bool)
-			results[i] = vv
-		}
-		return results
-	case time.Time:
-		results := make([]time.Time, len(raw))
-		for i, v := range raw {
-			vv, _ := v.(time.Time)
-			results[i] = vv
-		}
-		return results
-	default:
-		// TODO better error handling
-		log.DefaultLogger.Debug("Unsupported value %T", raw[0])
-		return nil
-	}
-}
